restructured routes and added controller files
