+++ server/utils/catchAsync.js
@@ -1,0 +1,5 @@
+module.exports = fn => {
+  return (req, res, next) => {
+    fn(req, res, next).catch(next);
+  };
+};
